Add methods to `ObjectFactory` to create `DirectoryProperty` and `RegularFileProperty` instances. These will replace the factory methods on `DefaultTask` and `ProjectLayout` to make construction of these types more consistent with other property types.
